# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CrisisConnect API
  description: |
    HTTP API for the **CrisisConnect** API gateway (`gateway-service.js`). All paths are relative to the gateway base URL.
    Most routes require a **Bearer** token from `POST /api/auth/login` (or `POST /api/auth/signup`), sent as `Authorization: Bearer <token>`.
  version: 1.0.0
  contact:
    name: CrisisConnect
servers:
  - url: http://localhost:4000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: hex
      description: Token returned from login or signup
  schemas:
...
    Resource:
      type: object
      properties:
        _id:
          type: string
        resourceType:
          type: string
          enum: [general, hospital_bed, blood_request]
        category:
          type: string
        quantity:
          type: integer
        location:
          type: string
        status:
          type: string
          enum: [Available, In Use, Depleted, Reserved]
        description:
          type: string
        wardType:
          type: string
        bedUnitLabel:
          type: string
        bloodGroup:
          type: string
        bloodUrgency:
          type: string
          enum: [low, medium, high, critical]
...
  parameters:
...
    MongoId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
paths:
...
  /api/resources:
    get:
      tags: [Resources]
      summary: List resources
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          description: Case-insensitive substring match on category name
          schema:
            type: string
        - name: status
          in: query
          description: Filter by availability status
          schema:
            type: string
            enum: [Available, In Use, Depleted, Reserved]
      responses:
        '200':
          description: Resource array
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Resource'
    post:
      tags: [Resources]
      summary: Create resource (staff/admin)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resource'
      responses:
        '201':
          description: Created
  /api/resources/{id}:
    put:
      tags: [Resources]
      summary: Update resource (staff/admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MongoId'
      responses:
        '200':
          description: Updated
    delete:
      tags: [Resources]
      summary: Delete resource (staff/admin)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/MongoId'
      responses:
        '200':
          description: Deleted
...